feat: nats user-course-created publisher listener
