+++ course_learning/src/events/listeners/queue-group-name.ts
@@ -1,0 +1,1 @@
+export const queueGroupName = 'course-learning-service';
